requirements.txt
PyJWT==2.1.0
requests==2.25.1